Add files via upload
Vehicles and its Data 
First of all we are going to do an analysis of a database on the whole vehicles. We will use some of the most important attributes in Python such as Pandas, Matloplib, Seaborn, etc. to show plots of multiple aspects of different events with the goal of displaying an unreadable database and turning it into something much friendlier.